Fixed up demo script to work on my system (ubuntu based) and without having to install libvterm anywhere.
Added a test script to make a 'final screen capture' of a sequence of
terminal codes.

Added library functions to access the 'frame buffer' without pty/fork/fd
artifacts - we don't always want or need all that stuff, sometimes we
just want to render a character stream.

Modified the library and the test program to be operable without
ncurses - the longer term plan is to make the library compilable without
curses, which will ease installation dependencies on some systems.